+++ src/test/java/org/jabref/logic/journals/AbbreviationsTest.java
@@ -4,56 +4,31 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
-public class AbbreviationsTest {
+class AbbreviationsTest {
 
-    private JournalAbbreviationPreferences prefs;
-    private JournalAbbreviationLoader abbreviations;
+    private JournalAbbreviationRepository repository;
 
     @BeforeEach
-    public void setUp() {
-        prefs = mock(JournalAbbreviationPreferences.class);
-        abbreviations = new JournalAbbreviationLoader();
+    void setUp() {
+        repository = new JournalAbbreviationRepository();
     }
 
     @Test
-    public void getNextAbbreviationAbbreviatesIEEEJournalTitle() {
-        when(prefs.useIEEEAbbreviations()).thenReturn(true);
-
-        assertEquals("#IEEE_J_PROC#",
-                abbreviations.getRepository(prefs)
-                             .getNextAbbreviation("Proceedings of the IEEE").get());
-    }
-
-    @Test
-    public void getNextAbbreviationExpandsIEEEAbbreviation() {
-        when(prefs.useIEEEAbbreviations()).thenReturn(true);
-
-        assertEquals("Proceedings of the IEEE",
-                abbreviations.getRepository(prefs)
-                             .getNextAbbreviation("#IEEE_J_PROC#").get());
-    }
-
-    @Test
-    public void getNextAbbreviationAbbreviatesJournalTitle() {
+    void getNextAbbreviationAbbreviatesJournalTitle() {
         assertEquals("Proc. IEEE",
-                abbreviations.getRepository(prefs)
-                             .getNextAbbreviation("Proceedings of the IEEE").get());
+                repository.getNextAbbreviation("Proceedings of the IEEE").get());
     }
 
     @Test
-    public void getNextAbbreviationRemovesPoint() {
+    void getNextAbbreviationRemovesPoint() {
         assertEquals("Proc IEEE",
-                abbreviations.getRepository(prefs)
-                             .getNextAbbreviation("Proc. IEEE").get());
+                repository.getNextAbbreviation("Proc. IEEE").get());
     }
 
     @Test
-    public void getNextAbbreviationExpandsAbbreviation() {
+    void getNextAbbreviationExpandsAbbreviation() {
         assertEquals("Proceedings of the IEEE",
-                abbreviations.getRepository(prefs)
-                             .getNextAbbreviation("Proc IEEE").get());
+                repository.getNextAbbreviation("Proc IEEE").get());
     }
 }